--- Rokéta/ConsoleObjectModules/ConsoleObjectSubclasses/EnemyModules/EnemyGenModules/IPosGenStrategy.cs
@@ -1,9 +1,0 @@
-﻿using Rokéta.ConsoleObjectModules.ConsoleObjectSubclasses.PlayerModules;
-
-namespace Rokéta.ConsoleObjectModules.ConsoleObjectSubclasses.EnemyModules.EnemyGenModules
-{
-    public interface IPosGenStrategy
-    {
-        int[] GetPos(Enemy enemy, Player player);
-    }
-}
