--- Rokéta/ConsoleObjectModules/ConsoleObjectSubclasses/PickupObjectModules/PickupObjectGenerator.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Rokéta.ConsoleObjectModules.ConsoleObjectSubclasses.PickupObjectModules
-{
-	internal class PickupObjectGenerator
-	{
-	}
-}
