+++ Rokéta/GameObjectModules/Encrypter.cs
